feature: add a git root keyword for `--config` argument
When operating in a monorepo the new functionality of tflint (`--chdir`) breaks config file path by changing the current working directory when tflint runs.  This commit patches this gap by allowing a user to provide __GIT_ROOT__ or populate a environment variable with a custom string and when the `--config` argument is is found it replaces that keyword with the repository root (pwd in the case of pre-commit).